git-svn-id: https://miau.svn.sourceforge.net/svnroot/miau/trunk@324 e7bde6ef-2016-0410-acfd-d3fa7a0704fe

+++ src/onconnect.c
@@ -1,6 +1,6 @@
 /* $Id$
  * -------------------------------------------------------
- * Copyright (C) 2003-2005 Tommi Saviranta <wnd@iki.fi>
+ * Copyright (C) 2003-2006 Tommi Saviranta <wnd@iki.fi>
  * -------------------------------------------------------
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -38,7 +38,7 @@ void
 onconnect_add(const char type, const char *target, const char *data)
 {
 	char *msg;
-	int mlen;
+	size_t mlen;
 
 	switch (type) {
 		case 'p':
